Pull add new NPC and Gear Item button component into a generic, event generating component.
