--- src/components/gear/add-new-gear-item-button/add-new-gear-item-button.html
@@ -1,8 +1,0 @@
-<button
-    class="text-white bg-black hover:bg-gray-900 active:bg-gray-700 font-bold px-3 py-3 rounded-full transition duration-150 ease-in-out cursor-pointer uppercase"
-    onclick="this.onAddNewGearItemButtonClick(event)"
->
-    <svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 -960 960 960" width="24px" fill="#e3e3e3">
-        <path d="M440-120v-320H120v-80h320v-320h80v320h320v80H520v320h-80Z" />
-    </svg>
-</button>
