+++ src/components/pages/gear/gear.component.html
@@ -1,4 +1,4 @@
 <error-panel></error-panel>
 <gear-list-filter-bar></gear-list-filter-bar>
 <gear-list></gear-list>
-<add-new-gear-item-button class="fixed bottom-20 right-5"></add-new-gear-item-button>
+<floating-add-new-entity-button></floating-add-new-entity-button>
