+++ src/components/pages/npcs/npcs.html
@@ -1,4 +1,4 @@
 <error-panel></error-panel>
 <npc-list-filter-bar></npc-list-filter-bar>
 <npc-list></npc-list>
-<add-new-npc-button class="fixed bottom-20 right-5"></add-new-npc-button>
+<floating-add-new-entity-button></floating-add-new-entity-button>
