+++ src/lib/events/add-new-entity-button-clicked.ts
@@ -1,0 +1,7 @@
+import { AppEvent } from "./app-event";
+
+export class AddNewEntityButtonClicked extends AppEvent {
+    constructor() {
+        super(AddNewEntityButtonClicked.name);
+    }
+}
